Re-Added delete project user dialog as a preprocessor

+++ SitecoreInstaller.App/Pipelines/Steps/Uninstall/DeleteProject.cs
@@ -1,23 +1,12 @@
-﻿using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Text;
-
-namespace SitecoreInstaller.App.Pipelines.Steps.Uninstall
+﻿namespace SitecoreInstaller.App.Pipelines.Steps.Uninstall
 {
-    using SitecoreInstaller.App.Pipelines.Preconditions;
-    using SitecoreInstaller.Domain.Pipelines;
-
-  public class DeleteProject : Step<PipelineEventArgs>
+  public class DeleteProject : Step<CleanupEventArgs>
+  {
+    protected override void InnerInvoke(object sender, CleanupEventArgs args)
     {
-        public DeleteProject()
-        {
-            AddPrecondition<CheckUserAcceptForDeleteProject>();
-        }
-
-        protected override void InnerInvoke(object sender, PipelineEventArgs args)
-        {
-            Services.Projects.DeleteProject(args.ProjectSettings.ProjectFolder.Directory);
-        }
+      //we only delete project if we're deep cleaning
+      if (args.DeepClean)
+        Services.Projects.DeleteProject(args.ProjectSettings.ProjectFolder.Directory);
     }
+  }
 }
